Get the basics working for server side rendering
