+++ bin/www
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-var app = require('../src/scripts/server'),
+var app = require('../src/server'),
     debug = require('debug')('www');
