+++ gruntfile.js
@@ -9,14 +9,17 @@ module.exports = function(grunt) {
     },
     jshint: {
       all: ['<%= paths.js %>'],
-      options: {jshintrc: true}
+      options: {
+        jshintrc: true,
+        additionalSuffixes: ['.js']
+      }
     },
     jscs: {
       all: ['<%= paths.js %>']
     }
   });
 
-  grunt.loadNpmTasks('grunt-contrib-jshint');
+  grunt.loadNpmTasks('grunt-jsxhint');
   grunt.loadNpmTasks('grunt-jscs');
 
   grunt.registerTask('default', []);
