+++ package.json
@@ -18,12 +18,16 @@
   "devDependencies": {
     "grunt": "^0.4.5",
     "grunt-cli": "^0.1.13",
-    "grunt-contrib-jshint": "^0.11.2",
-    "grunt-jscs": "^2.1.0"
+    "grunt-jscs": "^2.1.0",
+    "grunt-jsxhint": "^0.6.0",
+    "react-tools": "^0.13.3"
   },
   "dependencies": {
     "debug": "^2.2.0",
     "express": "^4.13.3",
-    "morgan": "^1.6.1"
+    "express-hbs": "^0.8.4",
+    "morgan": "^1.6.1",
+    "node-jsx": "^0.13.3",
+    "react": "^0.13.3"
   }
 }
