+++ src/server/.jshintrc
