+++ src/server/views/collection.hbs
@@ -1,0 +1,5 @@
+{{!< ./layouts/base }}
+
+{{#contentFor "head"}}
+<title>Collection</title>
+{{/contentFor}}
