+++ src/server/views/layouts/base.hbs
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    {{{block "head"}}}
+  </head>
+  <body>
+    {{{component}}}
+  </body>
+</html>
